+++ src/lib/components/table/title-table.vue
@@ -18,14 +18,15 @@
                         v-if="colSetMap[key] && colSetMap[key].name"
                         v-show="!colSetMap[key] || !colSetMap[key].hide"
                         :key="key"
+                        class="content-th"
                     >
                         <component
                             :is="{
                                 template : ('<span>' + colSetMap[key].name + '</span>'),
-                                components : vm.$root.$options.components,
+                                components : vm.$vnode.context.$options.components,
                                 data : function () {
                                     return {
-                                        context : this.$parent.$parent.$vnode.context
+                                        context : vm.$vnode.context
                                     };
                                 }
                             }"
@@ -61,6 +62,7 @@
                         v-else
                         v-show="!colSetMap[key] || !colSetMap[key].hide"
                         :key="key"
+                        class="content-th"
                     >
                         <span class="th-sort" v-if="colSetMap[key] && colSetMap[key].sort">
                             <i class="mo-icon mo-icon-sort no" @click="sortCol(key)" v-if="!data.sort[key] || (data.sort[key].type !== 'asc' && data.sort[key].type !== 'desc')"></i>
@@ -117,6 +119,7 @@
                     <td
                         v-show="(!colSetMap[data.titleKeys[index]] || !colSetMap[data.titleKeys[index]].hide) && (row._matchFilter === true || row._matchFilter === undefined)"
                         :key="index"
+                        class="content-td"
                     
                         @click="$emit('cell-click', line, data.normalKeys[index])"
                         @mouseenter="$emit('cell-enter', line, data.normalKeys[index])"
@@ -126,10 +129,10 @@
                         <component
                             :is="{
                                 template : ('<div>' + col + '</div>'),
-                                components : vm.$root.$options.components,
+                                components : vm.$vnode.context.$options.components,
                                 data : function () {
                                     return {
-                                        context : this.$parent.$parent.$vnode.context
+                                        context : vm.$vnode.context
                                     };
                                 }
                             }"